change: moving test files into their own directory.
I know this isn't really a best practice in go, but I have a strong preference for breaking out files into manageable chunks, which has resulted in many test files. To keep the functional branch uncluttered and the test files a bit more distinct, I'd like to use this sub-directory for the tests.